Move Sqrt expression to its own source file
